messaging_app/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
drf-nested-routers==0.94.2
sqlparse==0.5.3
typing_extensions==4.14.1
